move verifyToken() in it's own function
